refactor(export): extract OPML serialisation helpers into dedicated module
- exportService.ts was 649 lines; bundled OPML serialization, TaskPaper
  serialization, TaskPaper parsing, and the service dispatcher
- OPML helpers (xmlAttr, renderTaskOutline, renderProjectOutline) are
  pure functions with zero coupling to other concerns — now live in
  src/services/export/opml.ts where the OPML scope is self-evident
- Sets the pattern for follow-up issues to split TaskPaper render/parse

Closes #265
